+++ src/main/java/com/yahveh/dto/request/DetalleNotaEntregaRequest.java
@@ -4,13 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class DetalleNotaEntregaRequest {
     private String codArticulo;
     private int cantidad;
-    private float precioUnitario;
-    private float precioSinFactura;
-    private float descuento;
+    private BigDecimal precioUnitario;
+    private BigDecimal precioSinFactura;
+    private float descuento;          // porcentaje (0-100), no es dinero
 }
