+++ src/main/java/com/yahveh/dto/request/InventarioRequest.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -13,7 +14,7 @@ public class InventarioRequest {
     private String codArticulo;
     private String tipoMovimiento; // ENTRADA, SALIDA, AJUSTE, INVENTARIO_INICIAL
     private int cantidad;
-    private float precioUnitario;
+    private BigDecimal precioUnitario;
     private LocalDate fecha;
     private String observacion;
     private String codImportacion;
