+++ src/main/java/com/yahveh/dto/response/ArticuloResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -16,8 +18,8 @@ public class ArticuloResponse {
     private String descripcion;
     private String descripcion2;
     private int stockActual;
-    private double precioActual;
-    private double precioSinFactura;
+    private BigDecimal precioActual;
+    private BigDecimal precioSinFactura;
     private int audUsuario;
     /** Numero de fila global (1..N) calculado por el SP. Útil para enumerar incluso paginando. */
     private long rowNumber;
